get & set runtime vars from scripts
