+++ lib/editor/widgets/flow_editor_http/flow_editor_http.dart
@@ -10,6 +10,7 @@ import 'package:trayce/editor/models/header.dart';
 import 'package:trayce/editor/models/request.dart';
 import 'package:trayce/editor/repo/config_repo.dart';
 import 'package:trayce/editor/repo/explorer_service.dart';
+import 'package:trayce/editor/repo/runtime_vars_repo.dart';
 import 'package:trayce/editor/repo/send_request.dart';
 import 'package:trayce/editor/widgets/code_editor/code_editor_multi.dart';
 import 'package:trayce/editor/widgets/code_editor/url_input.dart';
@@ -190,6 +191,7 @@ class _FlowEditorHttpState extends State<FlowEditorHttp> with TickerProviderStat
 
     try {
       final explorerService = context.read<ExplorerService>();
+      final runtimeVarsRepo = context.read<RuntimeVarsRepo>();
       final config = context.read<ConfigRepo>().get();
       final httpClient = context.read<HttpClientI>();
       print("===============> FlowEditorHttp sendRequest ${_formRequest.url}");
@@ -200,6 +202,7 @@ class _FlowEditorHttpState extends State<FlowEditorHttp> with TickerProviderStat
             node: widget.node,
             collectionNode: widget.collectionNode,
             explorerService: explorerService,
+            runtimeVarsRepo: runtimeVarsRepo,
             config: config,
           ).send();
